requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-chroma
chromadb
google-generativeai
python-dotenv
fastapi
langserve
sse_starlette
uvicorn
pymupdf
sentence-transformers
bs4
faiss-cpu
tqdm
tiktoken
langchainhub
tavily-python
pytest
langgraph
langgraph-checkpoint-sqlite
